Fix code scanning alert #43: Multiplication result converted to larger type

+++ drivers/firmware/efi/memattr.c
@@ -39,7 +39,7 @@ int __init efi_memattr_init(void)
 		goto unmap;
 	}
 
-	tbl_size = sizeof(*tbl) + tbl->num_entries * tbl->desc_size;
+	tbl_size = sizeof(*tbl) + (unsigned long)tbl->num_entries * tbl->desc_size;
 	memblock_reserve(efi_mem_attr_table, tbl_size);
 	set_bit(EFI_MEM_ATTR, &efi.flags);
